+++ automation/src/main/java/org/apache/cloudberry/pxf/automation/applications/S3Application.java
@@ -19,139 +19,80 @@ package org.apache.cloudberry.pxf.automation.applications;
  * under the License.
  */
 
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.client.builder.AwsClientBuilder;
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import com.amazonaws.services.s3.model.DeleteObjectsRequest;
+import com.amazonaws.services.s3.model.ListObjectsV2Request;
+import com.amazonaws.services.s3.model.ListObjectsV2Result;
+import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.amazonaws.services.s3.model.S3ObjectSummary;
 import org.apache.cloudberry.pxf.automation.testcontainers.MinIOContainer;
-import org.apache.cloudberry.pxf.automation.testcontainers.PXFCloudberryContainer;
-import org.testcontainers.containers.Container.ExecResult;
 
 import java.io.IOException;
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
- * Runtime PXF server-config mutations needed by CloudAccess automation tests.
- *
- * The stable servers ({@code s3}, {@code s3-invalid}) are pre-baked into the
- * container image by entrypoint.sh. This class only handles things that must
- * change between tests: creating/removing the {@code s3-non-existent} server
- * with endpoint-only config, stripping or restoring the default server's
- * Hadoop XMLs, and clearing the gpadmin AWS credentials file.
+ * S3 API access wrapper used by automation tests to seed and clean fixtures
+ * in a MinIO bucket. Owns the AmazonS3 client; callers should call
+ * `shutdown()` when done (typically in afterClass before stopping the
+ * MinIO container).
  */
-public class S3Application {
+public class S3Application implements AutoCloseable {
 
-    private static final String SCRIPTS_PREFIX =
-            "/home/gpadmin/workspace/cloudberry-pxf/automation/src/main/resources/testcontainers/pxf-cbdb/script";
+    private final AmazonS3 s3Client;
 
-    private final PXFCloudberryContainer container;
-
-    public S3Application(PXFCloudberryContainer container) {
-        this.container = container;
+    public S3Application(MinIOContainer minio) {
+        this.s3Client = buildS3Client(minio.getHostEndpoint(), minio.getAccessKey(), minio.getSecretKey());
     }
 
-    // Writes s3-site.xml with endpoint only (no credentials) for credential-via-URL tests.
-    public void configureServerEndpointOnly(MinIOContainer minio, String serverName)
-            throws IOException, InterruptedException {
-        String endpoint = minio.getInternalEndpoint();
-
-        String script = String.join("\n",
-                "set -e",
-                "source " + SCRIPTS_PREFIX + "/pxf-env.sh",
-                "TEMPLATES_DIR=${PXF_HOME}/templates",
-                "PXF_BASE_SERVERS=${PXF_BASE}/servers",
-                "SERVER_DIR=${PXF_BASE_SERVERS}/" + serverName,
-                "mkdir -p \"${SERVER_DIR}\"",
-                "cp \"${TEMPLATES_DIR}/mapred-site.xml\" \"${SERVER_DIR}/\"",
-                "cat > \"${SERVER_DIR}/s3-site.xml\" <<'S3SITE'",
-                "<?xml version=\"1.0\" encoding=\"UTF-8\"?>",
-                "<configuration>",
-                "    <property>",
-                "        <name>fs.s3a.endpoint</name>",
-                "        <value>" + endpoint + "</value>",
-                "    </property>",
-                "    <property>",
-                "        <name>fs.s3a.path.style.access</name>",
-                "        <value>true</value>",
-                "    </property>",
-                "    <property>",
-                "        <name>fs.s3a.connection.ssl.enabled</name>",
-                "        <value>false</value>",
-                "    </property>",
-                "    <property>",
-                "        <name>fs.s3a.impl</name>",
-                "        <value>org.apache.hadoop.fs.s3a.S3AFileSystem</value>",
-                "    </property>",
-                "    <property>",
-                "        <name>fs.s3a.aws.credentials.provider</name>",
-                "        <value>org.apache.hadoop.fs.s3a.SimpleAWSCredentialsProvider</value>",
-                "    </property>",
-                "</configuration>",
-                "S3SITE"
-        );
-
-        ExecResult result = container.execInContainer("bash", "-l", "-c", script);
-        if (result.getExitCode() != 0) {
-            throw new RuntimeException(
-                    "Endpoint-only S3 server configuration failed (exit " + result.getExitCode() + "):\n"
-                            + result.getStdout() + "\n" + result.getStderr());
+    public void createBucket(String bucket) {
+        if (!s3Client.doesBucketExistV2(bucket)) {
+            s3Client.createBucket(bucket);
         }
-
-        new PXFApplication(container).restartPxf();
-        System.out.println("[S3Application] Endpoint-only PXF server '" + serverName + "' configured");
     }
 
-    public void clearGpadminAwsCredentials() throws IOException, InterruptedException {
-        runContainerScript("rm -f /home/gpadmin/.aws/credentials", "Cleared gpadmin AWS credentials file");
+    public void putObject(String bucket, String key, Path localFile) throws IOException {
+        s3Client.putObject(new PutObjectRequest(bucket, key, localFile.toFile()));
     }
 
-    public void removeServerDirectory(String serverName) throws IOException, InterruptedException {
-        runContainerScript(
-                "rm -rf \"${PXF_BASE}/servers/" + serverName + "\"",
-                "Removed PXF server directory '" + serverName + "'");
+    public void deletePrefix(String bucket, String prefix) {
+        ListObjectsV2Request request = new ListObjectsV2Request()
+                .withBucketName(bucket)
+                .withPrefix(prefix);
+        ListObjectsV2Result listing;
+        do {
+            listing = s3Client.listObjectsV2(request);
+            List<String> keys = new ArrayList<>();
+            for (S3ObjectSummary summary : listing.getObjectSummaries()) {
+                keys.add(summary.getKey());
+            }
+            if (!keys.isEmpty()) {
+                s3Client.deleteObjects(new DeleteObjectsRequest(bucket).withKeys(keys.toArray(new String[0])));
+            }
+            request.setContinuationToken(listing.getNextContinuationToken());
+        } while (listing.isTruncated());
     }
 
-    // Removes all Hadoop site files from the default PXF server (no HDFS cluster configured).
-    public void stripDefaultServerHdfsConfig() throws IOException, InterruptedException {
-        runDefaultServerScript(
-                "rm -f \"${SERVER_DIR}\"/hdfs-site.xml \"${SERVER_DIR}\"/mapred-site.xml"
-                        + " \"${SERVER_DIR}\"/yarn-site.xml \"${SERVER_DIR}\"/core-site.xml"
-                        + " \"${SERVER_DIR}\"/hbase-site.xml \"${SERVER_DIR}\"/hive-site.xml"
-                        + " \"${SERVER_DIR}\"/s3-site.xml",
-                "Stripped Hadoop/S3 site config from default PXF server");
-        clearGpadminAwsCredentials();
+    public void shutdown() {
+        s3Client.shutdown();
     }
 
-    // Restores Hadoop site files on the default PXF server from PXF templates.
-    public void restoreDefaultServerHdfsConfig() throws IOException, InterruptedException {
-        runDefaultServerScript(
-                "cp \"${TEMPLATES_DIR}\"/hdfs-site.xml \"${SERVER_DIR}/\""
-                        + " && cp \"${TEMPLATES_DIR}\"/mapred-site.xml \"${SERVER_DIR}/\""
-                        + " && cp \"${TEMPLATES_DIR}\"/yarn-site.xml \"${SERVER_DIR}/\""
-                        + " && cp \"${TEMPLATES_DIR}\"/core-site.xml \"${SERVER_DIR}/\""
-                        + " && cp \"${TEMPLATES_DIR}\"/hbase-site.xml \"${SERVER_DIR}/\""
-                        + " && cp \"${TEMPLATES_DIR}\"/hive-site.xml \"${SERVER_DIR}/\"",
-                "Restored Hadoop config on default PXF server");
+    @Override
+    public void close() {
+        shutdown();
     }
 
-    private void runDefaultServerScript(String serverDirAction, String logMessage)
-            throws IOException, InterruptedException {
-        String script = String.join("\n",
-                "set -e",
-                "source " + SCRIPTS_PREFIX + "/pxf-env.sh",
-                "TEMPLATES_DIR=${PXF_HOME}/templates",
-                "SERVER_DIR=${PXF_BASE}/servers/default",
-                serverDirAction
-        );
-        runContainerScript(script, logMessage);
-    }
-
-    private void runContainerScript(String body, String logMessage) throws IOException, InterruptedException {
-        ExecResult result = container.execInContainer("bash", "-l", "-c", body);
-        if (result.getExitCode() != 0) {
-            throw new RuntimeException(
-                    logMessage + " failed (exit " + result.getExitCode() + "):\n"
-                            + result.getStdout() + "\n" + result.getStderr());
-        }
-
-        if (body.contains("SERVER_DIR") || body.contains("servers/")) {
-            new PXFApplication(container).restartPxf();
-        }
-        System.out.println("[S3Application] " + logMessage);
+    private static AmazonS3 buildS3Client(String endpoint, String accessKey, String secretKey) {
+        return AmazonS3ClientBuilder.standard()
+                .withEndpointConfiguration(new AwsClientBuilder.EndpointConfiguration(endpoint, "us-east-1"))
+                .withPathStyleAccessEnabled(true)
+                .withCredentials(new AWSStaticCredentialsProvider(
+                        new BasicAWSCredentials(accessKey, secretKey)))
+                .build();
     }
 }
